:technologist: Hide private functions
